+++ resource_server/src/main/java/eu/europa/ec/eudi/signer/r3/resource_server/ResourceServerApplication.java
@@ -16,10 +16,7 @@
 
 package eu.europa.ec.eudi.signer.r3.resource_server;
 
-import eu.europa.ec.eudi.signer.r3.common_tools.utils.CryptoProperties;
-import eu.europa.ec.eudi.signer.r3.resource_server.config.AuthConfig;
-import eu.europa.ec.eudi.signer.r3.resource_server.config.CredentialsConfig;
-import eu.europa.ec.eudi.signer.r3.resource_server.config.InfoConfig;
+import eu.europa.ec.eudi.signer.r3.resource_server.config.*;
 import eu.europa.ec.eudi.signer.r3.resource_server.model.certificates.ejbca.EjbcaProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -27,7 +24,7 @@ import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
-@EnableConfigurationProperties({ InfoConfig.class, EjbcaProperties.class, AuthConfig.class, CredentialsConfig.class })
+@EnableConfigurationProperties({ InfoConfig.class, EjbcaProperties.class, AuthConfig.class, CredentialsConfig.class, KeysProperties.class, CertificatesProperties.class })
 @ComponentScan("eu.europa.ec.eudi.signer.r3")
 public class ResourceServerApplication {