+++ resource_server/src/main/java/eu/europa/ec/eudi/signer/r3/resource_server/model/CredentialsService.java
@@ -17,13 +17,12 @@
 package eu.europa.ec.eudi.signer.r3.resource_server.model;
 
 import eu.europa.ec.eudi.signer.r3.resource_server.model.certificates.CertificatesService;
-import eu.europa.ec.eudi.signer.r3.resource_server.model.certificates.ejbca.EjbcaService;
+import eu.europa.ec.eudi.signer.r3.resource_server.model.certificates.issuer.ICertificateIssuer;
 import eu.europa.ec.eudi.signer.r3.resource_server.model.database.entities.CertificateChain;
 import eu.europa.ec.eudi.signer.r3.resource_server.model.database.entities.Credentials;
 import eu.europa.ec.eudi.signer.r3.resource_server.model.database.repositories.CredentialsRepository;
+import eu.europa.ec.eudi.signer.r3.resource_server.model.keys.IKeysService;
 import eu.europa.ec.eudi.signer.r3.resource_server.model.keys.KeyPairRegister;
-import eu.europa.ec.eudi.signer.r3.resource_server.model.keys.KeysService;
-import eu.europa.ec.eudi.signer.r3.resource_server.model.keys.hsm.HsmService;
 import eu.europa.ec.eudi.signer.r3.resource_server.web.dto.CredentialsInfo.CredentialsInfoCert;
 import eu.europa.ec.eudi.signer.r3.resource_server.web.dto.CredentialsInfo.CredentialsInfoKey;
 import eu.europa.ec.eudi.signer.r3.resource_server.web.dto.CredentialsInfo.CredentialsInfoAuth;
@@ -39,17 +38,15 @@ import org.springframework.stereotype.Service;
 
 @Service
 public class CredentialsService {
-
     private final CertificatesService certificatesService;
     private final CredentialsRepository credentialsRepository;
-    private final KeysService keysService;
+    private final IKeysService keysService;
     private static final Logger logger = LoggerFactory.getLogger(CredentialsService.class);
 
-    public CredentialsService(@Autowired HsmService hsmService, @Autowired EjbcaService ejbcaService,
-          @Autowired CredentialsRepository credentialsRepository){
+    public CredentialsService(@Autowired IKeysService keysService, @Autowired ICertificateIssuer certificateIssuer, @Autowired CredentialsRepository credentialsRepository){
         this.credentialsRepository = credentialsRepository;
-        this.keysService = new KeysService(hsmService);
-        this.certificatesService = new CertificatesService(hsmService, ejbcaService);
+        this.keysService = keysService;
+        this.certificatesService = new CertificatesService(keysService, certificateIssuer);
     }
 
     /**