+++ resource_server/src/main/java/eu/europa/ec/eudi/signer/r3/resource_server/model/certificates/ejbca/EjbcaProperties.java
@@ -51,6 +51,10 @@ public class EjbcaProperties {
         this.countries = countries;
     }
 
+    public List<CountryConfig> getCountries() {
+        return countries;
+    }
+
     public String getCertificateAuthorityName(String country) {
         String def = null;
         boolean found = false;